requirements.txt
aiohttp==3.6.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
daiquiri==2.1.1
Deprecated==1.2.10
Django==3.0.5
django-crontab==0.7.1
django-filter==2.2.0
djangorestframework==3.11.0
git-pull-request==5.0.1
idna==2.9
Markdown==3.2.1
microfs==1.4.3
multidict==4.7.6
mysqlclient==1.4.6
protobuf
pseudo-microbit==1.8
PyGithub==1.51
PyJWT==1.7.1
pyminifier==2.1
PyMySQL==0.9.3
pyserial==3.4
python-json-logger==0.1.11
pytz==2020.1
requests==2.23.0
six==1.15.0
slackclient==2.6.2
soupsieve==2.0.1
sqlparse==0.3.1
uflash==1.3.0
urllib3==1.25.9
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.4.2
